requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.5
django-polymorphic==3.1.0
filelock==3.12.4
Pillow==10.1.0
platformdirs==3.10.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3